requirements.txt
kafka-python
minio
python-ffmpeg==2.0.4